+++ src/com/genir/aitweaks/features/autofire/AutofireAI.kt
@@ -4,11 +4,9 @@ import com.fs.starfarer.api.GameState
 import com.fs.starfarer.api.Global
 import com.fs.starfarer.api.combat.*
 import com.fs.starfarer.api.util.IntervalUtil
+import com.genir.aitweaks.features.lidar.LidarShipAI
 import com.genir.aitweaks.utils.attack.*
-import com.genir.aitweaks.utils.extensions.firingCycle
-import com.genir.aitweaks.utils.extensions.hasBestTargetLeading
-import com.genir.aitweaks.utils.extensions.timeToAttack
-import com.genir.aitweaks.utils.extensions.totalRange
+import com.genir.aitweaks.utils.extensions.*
 import com.genir.aitweaks.utils.firstShipAlongLineOfFire
 import com.genir.aitweaks.utils.rotateAroundPivot
 import org.lazywizard.lazylib.MathUtils
@@ -149,8 +147,12 @@ class AutofireAI(private val weapon: WeaponAPI) : AutofireAIPlugin {
         if (target == null) return null
 
         val intercept = intercept(weapon, AttackTarget(target!!), currentParams()) ?: return null
-        return if (weapon.slot.isTurret) intercept
-        else aimHardpoint(intercept)
+        return when {
+            // Lidar AI aims hardpoints correctly with the entire ship.
+            weapon.ship.hasAIType<LidarShipAI>() -> intercept
+            weapon.slot.isTurret -> intercept
+            else -> aimHardpoint(intercept)
+        }
     }
 
     /** get current weapon attack parameters */